+++ api/news.go
@@ -1,4 +1,4 @@
-package web
+package api
 
 import (
 	"encoding/base64"
