+++ models/models.go
@@ -29,51 +29,33 @@ func New(sqlDB *sql.DB) (Models, error) {
 // if err := initNews(); err != nil {
 // 		sql := `CREATE TABLE public.news (
 // 	id serial NOT NULL PRIMARY KEY,
-// 	header varchar(250),
+// 	header text,
 // 	text text,
-// 	imagepath varchar(250)
+// 	imagepath text
 // )`
-
-// _, err = sqlDB.Exec(sql)
-// 	return err
-// }
 
 // if err := initProduct(); err != nil {
 // 		sqlProduct := `CREATE TABLE public.product (
 // id serial NOT NULL PRIMARY KEY,
 //   price integer,
-//   name varchar(250),
+//   name text,
 //   description text
 // )`
-
-// _, err = sqlDB.Query(sqlProduct)
-// if err != nil {
-// 	return
-// }
 
 // sqlProductImages := `CREATE TABLE public.images (
 // id serial NOT NULL PRIMARY KEY,
 // product_id integer NOT NULL REFERENCES public.product(id) ON DELETE CASCADE,
-// name varchar(250)
+// name text
 // )`
 // _, err = sqlDB.Query(sqlProductImages)
-
-// 	return err
-// }
 
 // if err := initUser(); err != nil {
 // TODO: admin creaton belongs to fixtures
 // 			sql := `CREATE TABLE public.user (
 // 	id serial NOT NULL PRIMARY KEY,
-//     username varchar(250),
-//     password varchar(60)
+//     username text,
+//     password text
 // )`
-
-// 	_, err = sqlDB.Query(sql)
-// 	if err != nil {
-// 		clog.Errorf("%s", err)
-// 		return
-// 	}
 
 // 	// creating admin
 // 	var hash []byte
@@ -98,7 +80,7 @@ func New(sqlDB *sql.DB) (Models, error) {
 // }
 
 // if err := initAuth(); err != nil {
-// 		sql := `CREATE TABLE public.authentification (
+// 		sql := `CREATE TABLE public.authentication (
 // id serial NOT NULL PRIMARY KEY,
 //   user_id int NOT NULL REFERENCES public.user (id) ON DELETE CASCADE,
 //   valid_until timestamp,
