+++ web/public/favicon.ico
